+++ ChangeLog.md
@@ -3,6 +3,7 @@
 - cciss: fix parsing enclosure with no enclosure name [#84][]
 - megacli: actually report that cache is disabled. [#85][]
 - arcconf: accept 'Ready' as OK drive state. [#87][]
+- tw_cli: tweak VERIFYING state [#89][]
 
 ## 3.2.1 (2014-10-07)
 
@@ -131,3 +132,4 @@ NOTES:
 [#84]: https://github.com/glensc/nagios-plugin-check_raid/issues/84
 [#85]: https://github.com/glensc/nagios-plugin-check_raid/issues/85
 [#87]: https://github.com/glensc/nagios-plugin-check_raid/issues/87
+[#89]: https://github.com/glensc/nagios-plugin-check_raid/pull/89